backend/requirements.txt
flask
flask-cors
torch
torchvision
opencv-python
Pillow
numpy
scikit-learn
ultralytics
psycopg2-binary
bcrypt
PyJWT
werkzeug
python-dotenv